+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- `services.getAuthCredentials()` resource implementation with tests and usage example
 - `services.getProlongations()` resource implementation with tests and usage example
 
 ## [0.3.1] - 2026-04-09